requirements.txt
fastapi
uvicorn
twilio
python-dotenv
requests
python-multipart
google-generativeai